Fix genericArgumentFactories w/  T? in field usage (#840)
When a generic type (`T`) is used in a field as nullable (`final T? someField;`) then we
cannot call the defined converters which are not expected to return `null` values (in `fromJson`)
or handle `null` argument (in `toJson`). So we do the `null` check in a helper and pass it through.

Fixes https://github.com/google/json_serializable.dart/issues/803 (for real)
Closes https://github.com/google/json_serializable.dart/pull/837

Co-authored-by: Nate Bosch <nbosch@google.com>